requirements.txt
kafka-python
psycopg2-binary
python-dotenv
streamlit
pandas
pyspark
